# docs/developer/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Fyredocs API
  description: |
    Fyredocs is a document processing platform that provides PDF conversion,
    optimization, and organization tools. All responses follow a unified
    envelope format (`APIResponse`) with consistent error handling.

    ## Authentication
    The API supports three authentication methods:
    - **BearerAuth** - JWT access token in the `Authorization: Bearer <token>` header (API clients).
    - **CookieAuth** - `access_token` HttpOnly cookie set on signup/login. Pair with `refresh_token` (Path=`/auth`) for silent rotation via `POST /auth/refresh`.
    - **GuestToken** - `guest_token` HttpOnly cookie issued automatically by the API gateway on first contact when no auth is present. The same cookie is sent on every subsequent request.

    Authenticated users get persistent job storage. Guest users get temporary
    jobs that expire after a configurable TTL.

    ## Presigned Multipart Uploads
    File bytes are uploaded directly to object storage (MinIO/S3) through
    presigned URLs — they never pass through the application services:
    1. `POST /api/upload/init` creates an S3 multipart upload and returns a
       presigned PUT URL for every part.
    2. The browser `PUT`s each part to its presigned URL and collects the
       returned `ETag` response headers.
    3. `GET /api/upload/{uploadId}/parts` re-presigns part URLs to resume an
       interrupted upload or refresh expired URLs.
...
       caller's plan limit.
...

...

...
servers:
...
paths:
...
  /auth/plans:
    get:
      tags: [Authentication]
      summary: List subscription plans
      description: |
        Returns all non-anonymous subscription plans with their file size and
        file count limits. No authentication required.
      operationId: getPlans
      responses:
        "200":
          description: Plans retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          plans:
                            type: array
                            items:
                              $ref: "#/components/schemas/SubscriptionPlan"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Creates a new user account and returns a JWT access token.
        The token is also set as an HTTP-only cookie.
        Password must be 8-128 characters.
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Registration successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: access_token=eyJ...; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      error:
                        $ref: "#/components/schemas/APIError"
                        example:
                          code: USER_ALREADY_EXISTS
                          details: User already exists
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/login:
    post:
      tags: [Authentication]
      summary: Log in with email and password
      description: |
        Authenticates a user and returns a JWT access token.
        The token is also set as an HTTP-only cookie.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: access_token=eyJ...; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: |
        Issues a new short-lived `access_token` cookie from a valid `refresh_token`
        cookie. The refresh token cookie is left unchanged — only the underlying
        `user_sessions` row's access-token hash is rotated. Rate-limited at
        `RATE_LIMIT_REFRESH` (default 10) per `RATE_LIMIT_WINDOW` (default 60s)
        per IP.
      operationId: refreshToken
      security: []  # auth comes from the refresh_token cookie itself
      responses:
        "200":
          description: New access token issued
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "access_token=eyJhbGc...; HttpOnly; Secure; SameSite=Lax; Max-Age=28800; Path=/"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: "#/components/schemas/User"
                          accessExpiresAt:
                            type: integer
                            format: int64
                            description: Access token expiry as unix milliseconds
        "401":
          description: Missing, expired, or unknown refresh token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      error:
                        $ref: "#/components/schemas/APIError"
                        example:
                          code: INVALID_REFRESH_TOKEN
                          details: Your session has expired. Please log in again.
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/dashboard:
    get:
      tags: [Admin Analytics]
      summary: Unified role-aware dashboard
      description: |
        Single landing endpoint for every authenticated user. The server filters
        the payload by role (from the gateway-set `X-User-Role`): `admin` /
        `super-admin` receive `data.role:"admin"` with system KPIs, regular users
        receive `data.role:"user"` with personal KPIs scoped to their own
        `user_id`. Guests are rejected. Computed from analytics-service's own
        event store — there is no separate `/admin/dashboard`.
      operationId: getDashboard
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Look-back window in days
      responses:
        "200":
          description: Role-filtered dashboard payload (data.role is "admin" or "user")
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Guests do not have a dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
...
  /api/documents:
    get:
      tags: [Documents]
      summary: List documents
      operationId: listDocuments
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [uploaded, processing, ready, failed] } }
        - { name: folderId, in: query, schema: { type: string, format: uuid } }
        - { name: tagId, in: query, schema: { type: string, format: uuid } }
        - { name: q, in: query, description: Full-text search (name + extracted content), schema: { type: string } }
        - { name: trashed, in: query, description: "true returns the Trash view", schema: { type: boolean } }
        - { name: orgId, in: query, description: Org-scope the listing (viewer+), schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: Paginated documents (data[] + meta{page,limit,total})
          content: { application/json: { schema: { $ref: "#/components/schemas/APIResponse" } } }
    post:
      tags: [Documents]
      summary: Create a document (metadata only)
      operationId: createDocument
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                folderId: { type: string, format: uuid }
                organizationId: { type: string, format: uuid, description: "Org-scope (editor+)" }
                fileType: { type: string }
                mimeType: { type: string }
                fileSize: { type: integer, format: int64 }
                storagePath: { type: string }
                status: { type: string }
                metadata: { type: object }
      responses:
        "201": { description: Document created, content: { application/json: { schema: { $ref: "#/components/schemas/APIResponse" } } } }
        "400": { description: Validation error }
...
  /api/folders:
    get:
      tags: [Documents]
      summary: List folders
      operationId: listFolders
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters: [{ name: parentId, in: query, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Folders } }
    post:
      tags: [Documents]
      summary: Create a folder
      operationId: createFolder
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string }, parentId: { type: string, format: uuid } } } } }
      responses: { "201": { description: Folder created } }
...
  /api/tags:
    get:
      tags: [Documents]
      summary: List tags
      operationId: listTags
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses: { "200": { description: Tags } }
    post:
      tags: [Documents]
      summary: Create a tag (idempotent on user+name)
      operationId: createTag
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string }, color: { type: string } } } } }
      responses: { "201": { description: Tag created } }
...
  /api/exports:
    get:
      tags: [Documents]
      summary: List exports
      operationId: listExports
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses: { "200": { description: Exports (status/metadata) } }
    post:
      tags: [Documents]
      summary: Queue an async export of documents in scope
      operationId: createExport
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format]
              properties:
                format: { type: string, enum: [csv, json] }
                organizationId: { type: string, format: uuid }
                status: { type: string }
                folderId: { type: string, format: uuid }
                tagId: { type: string, format: uuid }
      responses: { "201": { description: Export queued } }
...
  /api/orgs:
    get:
      tags: [Organizations]
      summary: List organizations the caller belongs to
      operationId: listOrganizations
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses: { "200": { description: "Orgs, each with the caller's role" } }
    post:
      tags: [Organizations]
      summary: Create an organization (caller becomes owner)
      operationId: createOrganization
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string } } } } }
      responses:
        "201": { description: Organization created }
        "400": { description: Invalid name }
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Recent notifications + unread count
      operationId: listNotifications
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200": { description: "{notifications:[max 50], unreadCount}", content: { application/json: { schema: { $ref: "#/components/schemas/APIResponse" } } } }
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token passed in the `Authorization: Bearer <token>` header.
        Tokens are issued on signup/login and have a configurable TTL (default 8h).

    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT access token stored in an HTTP-only, Secure, SameSite=Lax cookie.
        Set automatically on successful signup or login.

    GuestToken:
      type: apiKey
      in: header
      name: X-Guest-Token
      description: |
        Temporary guest identifier for unauthenticated users. Can also be
        sent via the `guest_token` cookie. Allows guests to track their
        processing jobs without creating an account.

  # ---------------------------------------------------------------------------
  # Reusable Response Headers
  # ---------------------------------------------------------------------------
...
  schemas:
...
    APIResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        message:
          type: string
          description: Human-readable message describing the outcome
        data:
          description: Response payload (type varies per endpoint; null on errors)
        error:
          nullable: true
          $ref: "#/components/schemas/APIError"
        meta:
          nullable: true
          $ref: "#/components/schemas/Meta"
    APIError:
      type: object
      required: [code, details]
      properties:
        code:
          type: string
          description: Machine-readable error code (e.g. INVALID_INPUT, NOT_FOUND)
          example: INVALID_INPUT
        details:
          type: string
          description: Human-readable error description
          example: Invalid request body
...
    SignupRequest:
      type: object
      required: [email, password, fullName, country]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: "secureP@ss123"
        fullName:
          type: string
          example: John Doe
        phone:
          type: string
          description: Optional phone number
          example: "+1234567890"
        country:
          type: string
          example: US
        image:
          type: string
          format: uri
          description: Optional profile image URL
          example: "https://example.com/avatar.jpg"
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: user@example.com
        fullName:
          type: string
          example: John Doe
        phone:
          type: string
          example: "+1234567890"
        country:
          type: string
          example: US
        image:
          type: string
          example: "https://example.com/avatar.jpg"
        role:
          type: string
          enum: [user, super-admin]
          example: user
          description: User role from database. 'super-admin' grants access to admin dashboard.
        planName:
          type: string
          enum: [free, pro]
          example: free
          description: The user's current subscription plan name

    # -- Subscription Plan Schema --
    SubscriptionPlan:
      type: object
      properties:
        name:
          type: string
          example: free
        maxFileSizeMb:
          type: integer
          example: 25
          description: Maximum file size in megabytes
        maxFilesPerJob:
          type: integer
          example: 10
          description: Maximum number of files per job
        retentionDays:
          type: integer
          example: 7
          description: Number of days job files are retained

    # -- Upload Schemas --
...
  responses:
...
    InternalError:
      description: Internal server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/APIResponse"
              - type: object
                properties:
                  success:
                    example: false
                  error:
                    $ref: "#/components/schemas/APIError"
                    example:
                      code: SERVER_ERROR
                      details: An unexpected error occurred